refactor(go): nesting indexers to pkg
