+++ controllers/externalclusterreference_controller.go
@@ -22,8 +22,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"github.com/clastix/cluster-api-control-plane-provider-kamaji/api/v1alpha1"
-	"github.com/clastix/cluster-api-control-plane-provider-kamaji/indexers"
 	"github.com/clastix/cluster-api-control-plane-provider-kamaji/pkg/externalclusterreference"
+	"github.com/clastix/cluster-api-control-plane-provider-kamaji/pkg/indexers"
 )
 
 type ExternalClusterReferenceReconciler struct {
