+++ pkg/indexers/setup.go
